Move net.sf.jabref.logic.bibtex package to net.sf.jabref.bibtex
