+++ src/main/java/net/sf/jabref/bibtex/EntrySorter.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.logic.bibtex;
+package net.sf.jabref.bibtex;
 
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.database.DatabaseChangeEvent;
