+++ src/main/java/net/sf/jabref/bibtex/comparator/BibtexStringComparator.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.logic.bibtex.comparator;
+package net.sf.jabref.bibtex.comparator;
 
 import net.sf.jabref.model.entry.BibtexString;
