+++ src/main/java/net/sf/jabref/collab/ChangeScanner.java
@@ -35,9 +35,9 @@ import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.importer.OpenDatabaseAction;
 import net.sf.jabref.importer.ParserResult;
-import net.sf.jabref.logic.bibtex.DuplicateCheck;
-import net.sf.jabref.logic.bibtex.EntrySorter;
-import net.sf.jabref.logic.bibtex.comparator.EntryComparator;
+import net.sf.jabref.bibtex.DuplicateCheck;
+import net.sf.jabref.bibtex.EntrySorter;
+import net.sf.jabref.bibtex.comparator.EntryComparator;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.database.BibtexDatabase;
 import net.sf.jabref.model.entry.BibtexEntry;
