+++ src/main/java/net/sf/jabref/gui/DuplicateResolverDialog.java
@@ -23,7 +23,7 @@ import java.io.StringWriter;
 
 import javax.swing.*;
 
-import net.sf.jabref.logic.bibtex.BibtexEntryWriter;
+import net.sf.jabref.bibtex.BibtexEntryWriter;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
