+++ src/main/java/net/sf/jabref/gui/StringDialog.java
@@ -53,7 +53,7 @@ import net.sf.jabref.model.database.KeyCollisionException;
 import net.sf.jabref.gui.undo.UndoableInsertString;
 import net.sf.jabref.gui.undo.UndoableRemoveString;
 import net.sf.jabref.gui.undo.UndoableStringChange;
-import net.sf.jabref.logic.bibtex.comparator.BibtexStringComparator;
+import net.sf.jabref.bibtex.comparator.BibtexStringComparator;
 import net.sf.jabref.logic.id.IdGenerator;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.database.BibtexDatabase;
