+++ src/main/java/net/sf/jabref/gui/TransferableBibtexEntry.java
@@ -16,7 +16,7 @@
 package net.sf.jabref.gui;
 
 import net.sf.jabref.exporter.LatexFieldFormatter;
-import net.sf.jabref.logic.bibtex.BibtexEntryWriter;
+import net.sf.jabref.bibtex.BibtexEntryWriter;
 import net.sf.jabref.model.entry.BibtexEntry;
 
 import javax.swing.*;
