+++ src/test/java/net/sf/jabref/bibtex/DuplicateCheckTest.java
@@ -1,8 +1,7 @@
-package net.sf.jabref.logic.bibtex;
+package net.sf.jabref.bibtex;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.logic.bibtex.DuplicateCheck;
 import net.sf.jabref.logic.id.IdGenerator;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.entry.BibtexEntryTypes;
